check argv INT_MAX

+++ libft/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yejeong <yejeong@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jeong-yena <jeong-yena@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/11 17:36:23 by yejeong           #+#    #+#             */
-/*   Updated: 2021/05/16 19:31:55 by yejeong          ###   ########.fr       */
+/*   Updated: 2022/02/03 18:30:17 by jeong-yena       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ static unsigned long	atoi_sub(const char *str)
 	return (rt);
 }
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int				i;
 	int				sum;
@@ -50,5 +50,5 @@ int	ft_atoi(const char *str)
 		return (-1);
 	else if (sum == -1 && rt > 9223372036854775807)
 		return (0);
-	return ((int)rt * sum);
+	return (rt * sum);
 }
