+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jeong-yena <jeong-yena@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/10 16:02:35 by kyunkim           #+#    #+#             */
-/*   Updated: 2022/01/28 19:50:34 by jeong-yena       ###   ########.fr       */
+/*   Updated: 2022/02/03 18:30:26 by jeong-yena       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ char				*ft_strrchr(const char *str, int ch);
 char				*ft_strnstr(const char *str, const char *to_find, size_t n);
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 int					ft_strcmp(const char *s1, const char *s2);
-int					ft_atoi(const char *str);
+long long			ft_atoi(const char *str);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
